# docs/contracts/openapi_2026-04-24.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SME Cyber Exposure Dashboard",
    "description": "AI-driven DAST platform for SMEs",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/scans/": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Create Scan",
        "description": "Create a new scan and start it in the background.\nSupports both legacy (target_url) and new (target_id) formats.",
        "operationId": "create_scan_api_v1_scans__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "List Scans",
        "description": "List all scans with optional status filter.",
        "operationId": "list_scans_api_v1_scans__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanSummary"
                  },
                  "title": "Response List Scans Api V1 Scans  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/schedule": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Create Schedule",
        "description": "Create a recurring scan schedule.",
        "operationId": "create_schedule_api_v1_scans_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/scans/schedules": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "List Schedules",
        "description": "List active scan schedules.",
        "operationId": "list_schedules_api_v1_scans_schedules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/scans/schedules/{schedule_id}": {
      "delete": {
        "tags": [
          "scans"
        ],
        "summary": "Delete Schedule",
        "description": "Cancel a scheduled scan.",
        "operationId": "delete_schedule_api_v1_scans_schedules__schedule_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/ai": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Create Ai Scan",
        "description": "Create a new AI-powered scan and enqueue it through Celery.\n\nReturns HTTP 202 Accepted \u2014 the scan is queued, not yet running.\nThe Celery worker runs the full AgentOrchestrator pipeline:\n1. ReconAgent -> 2. AttackAgent (Nuclei) -> 3. ValidationAgent -> 4. ReportingAgent\n\nPhase 2.1 hardening: previously used FastAPI BackgroundTasks, which died\non backend restart. Celery handles retries and survives restarts.",
        "operationId": "create_ai_scan_api_v1_scans_ai_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/scans/{scan_id}": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get Scan",
        "description": "Get detailed scan information including vulnerabilities and agent logs.",
        "operationId": "get_scan_api_v1_scans__scan_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "scans"
        ],
        "summary": "Delete Scan",
        "description": "Delete a scan and all its associated data.",
        "operationId": "delete_scan_api_v1_scans__scan_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/{scan_id}/logs": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get Scan Logs",
        "description": "Get AI agent logs for a specific scan.\nShows the reasoning chain of all agents.",
        "operationId": "get_scan_logs_api_v1_scans__scan_id__logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentLogResponse"
                  },
                  "title": "Response Get Scan Logs Api V1 Scans  Scan Id  Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/{scan_id}/stop": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Stop Scan",
        "description": "Stop a running scan.",
        "operationId": "stop_scan_api_v1_scans__scan_id__stop_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScanCreate": {
        "properties": {
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "target_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Url"
          },
          "scan_type": {
            "type": "string",
            "title": "Scan Type",
            "default": "full"
          },
          "tools": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tools"
          },
          "configuration": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration"
          },
          "schedule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule"
          },
          "auto_report": {
            "type": "boolean",
            "title": "Auto Report",
            "default": false
          },
          "siem_forward": {
            "type": "boolean",
            "title": "Siem Forward",
            "default": false
          }
        },
        "type": "object",
        "title": "ScanCreate"
      },
      "ScanDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          },
          "scan_type": {
            "type": "string",
            "title": "Scan Type"
          },
          "target_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Url"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score",
            "default": 0.0
          },
          "target_display": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Display",
            "default": "Unknown"
          },
          "vulnerabilities_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vulnerabilities Count",
            "default": 0
          },
          "assets_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assets Count",
            "default": 0
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "configuration": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration"
          },
          "agent_thoughts": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Thoughts"
          },
          "vulnerabilities": {
            "items": {
              "$ref": "#/components/schemas/VulnerabilityResponse"
            },
            "type": "array",
            "title": "Vulnerabilities",
            "default": []
          },
          "agent_logs": {
            "items": {
              "$ref": "#/components/schemas/AgentLogResponse"
            },
            "type": "array",
            "title": "Agent Logs",
            "default": []
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/ScanAssetResponse"
            },
            "type": "array",
            "title": "Assets",
            "default": []
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__scan__ActionItemResponse"
            },
            "type": "array",
            "title": "Actions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "scan_type"
        ],
        "title": "ScanDetail"
      },
      "ScanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          },
          "scan_type": {
            "type": "string",
            "title": "Scan Type"
          },
          "target_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Url"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score",
            "default": 0.0
          },
          "target_display": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Display",
            "default": "Unknown"
          },
          "vulnerabilities_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vulnerabilities Count",
            "default": 0
          },
          "assets_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assets Count",
            "default": 0
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "configuration": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "scan_type"
        ],
        "title": "ScanResponse"
      },
      "ScanStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "ScanStatus"
      },
      "ScanSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          },
          "scan_type": {
            "type": "string",
            "title": "Scan Type"
          },
          "target_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Url"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score",
            "default": 0.0
          },
          "target_display": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Display",
            "default": "Unknown"
          },
          "vulnerabilities_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vulnerabilities Count",
            "default": 0
          },
          "assets_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assets Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "scan_type"
        ],
        "title": "ScanSummary"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}